refactor(orgs): split org-data.ts (1822L, 50 exports) into per-record _data/ modules
Splits apps/web/src/app/organizations/[slug]/org-data.ts into 14 focused
files under _data/ (common, entity, grants, divisions, funding-programs,
personnel, funding-rounds, investments, equity-positions, dilution-stages,
charitable-pledges, related-orgs, resources, charts, loaders). The original
file becomes a 24-line barrel re-export so all 20 callers (page.tsx,
section components, legislation/investments/funding-rounds pages, resource
components, tests) keep their existing import paths unchanged.

Pure refactor — no behavior change. All 1688 tests pass, full pnpm build
succeeds, 54 gate checks pass, tsc --noEmit clean.

The underscore prefix on _data/ prevents Next.js from treating it as a
route segment (the existing /organizations/[slug]/data/ route would
otherwise collide).

Closes QUA-682
