+++ apps/web/src/app/organizations/[slug]/_data/related-orgs.ts
@@ -1,0 +1,7 @@
+export interface RelatedOrg {
+  id: string;
+  name: string;
+  slug: string | null;
+  relationship: string;
+  date: string | null;
+}
